moved single list functionality into new file
